Circuit List.
Users can now select the circuit button which show them a list of circuits via there name. When a user selects a circuit they are shown a list of years which represents the results from that year on that circuit. When they select a year they are shown the results for that year on that circuit.

Homescreen has had a header added shows displays the F1 logo along with "F1 App"